refactor/move e2e test merging
